RAML cleanup and addition of instance level hold management
The RAML was a bit sloppy and had some issues when viewing the API doc generated from it, so that was cleaned up. Also added API endpoints for instance level holds, which is currently not supported in FOLIO, but discovery services use instance (title) level holds a majority of the time since the patron doesn't typically care which copy of a title they get. This will allow discovery services to make the API calls, but get a reasonable error until this is supported.