Add mod-feesfines support for patron charges
In order to populate patron charges for discovery services, we need info from mod-feesfines. This module is pre-release and some of the functionality implemented in mod-patron makes assumptions about how the mod-feesfines APIs are intended to be used. These assumptions may turn out to be incorrect and we may need to revisit this once there is demonstrable data in the DB for mod-feesfines as well as working UI integration.