MODPATRON-8: integrate with circulation
Integrated the `patron` interface with the instance level hold API in `circulation`. Made some adjustments to the schema:

- We no longer accept the `fulfillmentPreference` property
- We now require the `pickupLocationId` property

Additionally, noticed logging was a bit of a mess since the module was updated to the latest RMB. Added log4j2 to fix this. Added the `requestDate` to the response, since it was missing (this likely fixes MODPATRON-20).