Use checkpoints in the unit tests
When we upgraded to JUnit 5, we used `context.completing()` to signal both when the verticle was up and when the mock Okapi server was up. In this case, when the first one completes, the setup is complete and testing begins. This could lead to unpredictable failures as the test could begin before the verticle is ready or before the mock okapi is ready. This appears to be what has happened in the latest branch build.

We now use a `Checkpoint`, which is similar to a `CountDownLatch`, to control when the setup finishes. In this case, we have 2 checkpoints. When each operation completes, we `flag()` the checkpoint representing the completed task, either verticle deployment or mock Okapi startup. Once each checkpoint issues their `flag()`, the test context is now "complete" and we can now execute a test.