MODPATRON-2: add request queue position to the hold (request)
When returning the request details in a patron account request or hold request, we now return the request queue position.

Also, the "requestLength" JSON field was removed as it is not necessary at this time and probably never will be.

Finally, altered the FOLIO request query to return all requests, not just holds. This way, the client will be able to see page and recalls as well, though the client (a discovery service) will not be able to distinguish between the request types.