MODPATRON-56: Inventory Optimistic Locking
mod-patron now can also return HTTP 409 status code for
optimistic locking failure. This requires a minor version
bump for patron interface.

mod-patron only uses selected properties and therefore is
not affected by inventory optimistic locking `_version`
property. The new inventory 11.0 interface can safely
been added as an alternative requirement.

Fetching the holdings record of the item is not needed,
this was used to get the instance of the item. This is
replaced by using this instance query:
`holdingsRecords.id==...`
This is faster, requires less code, and avoids the
(explicit) holdings-storage dependency.